Rename audio-event-provenance to kebab-case
Fixes: #2475
